refactor chat_apis
